+++ composer/core/logging/base_backend.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, List, Optional, Tuple
 
 from composer.core.callback import Callback, RankZeroCallback
 from composer.core.logging.logger import Logger
-from composer.utils.ddp import is_rank_zero
+from composer.utils.ddp import get_global_rank
 
 if TYPE_CHECKING:
     from composer.core.logging.logger import LogLevel, TLogData
@@ -116,7 +116,7 @@ class RankZeroLoggerBackend(BaseLoggerBackend, RankZeroCallback, ABC):
 
     @final
     def will_log(self, state: State, log_level: LogLevel) -> bool:
-        if not state.is_rank_zero:
+        if get_global_rank() != 0:
             return False
         return self._will_log(state, log_level)
 
@@ -138,7 +138,7 @@ class RankZeroLoggerBackend(BaseLoggerBackend, RankZeroCallback, ABC):
 
     @final
     def log_metric(self, epoch: int, step: int, log_level: LogLevel, data: TLogData) -> None:
-        if not is_rank_zero():
+        if get_global_rank() != 0:
             # no log if not on rank zero, clear deferred calls to free memory
             self._deferred_log_metric_calls = None
             return
