+++ composer/trainer/__init__.py
@@ -1,7 +1,6 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.trainer import devices as devices
-from composer.trainer.ddp import DDPDataLoader as DDPDataLoader
 from composer.trainer.trainer import Trainer as Trainer
 from composer.trainer.trainer_hparams import TrainerHparams as TrainerHparams
