+++ composer/utils/__init__.py
@@ -1,11 +1,7 @@
 # Copyright 2021 MosaicML. All Rights Reserved.
 
 from composer.utils import augmentation_primitives as augmentation_primitives
-from composer.utils.ddp import get_global_rank as get_global_rank
-from composer.utils.ddp import get_local_rank as get_local_rank
-from composer.utils.ddp import get_local_world_size as get_local_world_size
-from composer.utils.ddp import get_world_size as get_world_size
-from composer.utils.ddp import is_rank_zero as is_rank_zero
+from composer.utils import ddp as ddp
 from composer.utils.determinism import get_random_seed as get_random_seed
 from composer.utils.determinism import seed_all as seed_all
 from composer.utils.iter_helpers import ensure_tuple as ensure_tuple
