+++ tests/trainer/test_checkpoint.py
@@ -19,7 +19,7 @@ from composer.core.types import StateDict
 from composer.trainer.devices import CPUDeviceHparams, DeviceHparams, GPUDeviceHparams
 from composer.trainer.trainer import Trainer
 from composer.trainer.trainer_hparams import TrainerHparams, callback_registry
-from composer.utils.ddp import is_rank_zero
+from composer.utils import ddp
 from tests.test_state import assert_state_equivalent
 from tests.utils.deep_compare import deep_compare
 
@@ -175,7 +175,7 @@ def test_checkpoint(
     checkpoint_c_file_path = os.path.join(checkpoint_b_folder, final_checkpoint)
     trainer_2_hparams_filepath = os.path.join(checkpoint_b_folder, "hparams.yaml")
 
-    if is_rank_zero():
+    if ddp.get_global_rank() == 0:
 
         assert_checkpoints_equivalent(
             hparams_file_a=trainer_1_hparams_filepath,
