+++ tests/trainer/test_ddp.py
@@ -21,6 +21,7 @@ from composer.datasets import DataloaderHparams, DataloaderSpec, MemoryFormat, S
 from composer.trainer.devices import CPUDeviceHparams, GPUDeviceHparams
 from composer.trainer.devices.device_hparams import DeviceHparams
 from composer.trainer.trainer_hparams import TrainerHparams, callback_registry, dataset_registry
+from composer.utils import ddp
 from tests.fixtures.models import SimpleBatchPairModelHparams
 
 
@@ -228,7 +229,7 @@ def test_ddp(device: DeviceHparams, world_size: int, ddp_tmpdir: str, mosaic_tra
     is_train_to_pickles: Dict[bool, List[Dict[str, types.Tensor]]] = {True: [], False: []}
 
     for epoch in range(num_epochs):
-        for local_rank in range(trainer.state.local_world_size):
+        for local_rank in range(ddp.get_local_world_size()):
             for is_train in (True, False):
                 data: Dict[str, types.Tensor] = torch.load(  # type: ignore
                     get_batch_file_path(ddp_tmpdir, rank=local_rank, epoch=epoch, is_train=is_train),
