+++ tests/trainer/test_trainer.py
@@ -87,7 +87,7 @@ def test_trainer_determinism(mosaic_trainer_hparams: TrainerHparams):
     first_model = first_trainer.state.model.module
     assert isinstance(first_model, BaseMosaicModel)
     assert first_trainer.state.train_dataloader is not None
-    first_loss = get_total_loss(first_model, first_trainer.state.train_dataloader, first_trainer.ddp)
+    first_loss = get_total_loss(first_model, first_trainer.state.train_dataloader)
 
     # Second trainer must be created after fitting the first so that the
     # seeds get fully reset for the second training run
@@ -96,7 +96,7 @@ def test_trainer_determinism(mosaic_trainer_hparams: TrainerHparams):
     second_model = second_trainer.state.model.module
     assert isinstance(second_model, BaseMosaicModel)
     assert second_trainer.state.train_dataloader is not None
-    second_loss = get_total_loss(second_model, second_trainer.state.train_dataloader, second_trainer.ddp)
+    second_loss = get_total_loss(second_model, second_trainer.state.train_dataloader)
 
     torch.testing.assert_allclose(second_loss, first_loss)
