+++ tests/utils/trainer_fit.py
@@ -9,14 +9,13 @@ from composer.datasets.synthetic import SyntheticDataLabelType, SyntheticDataset
 from composer.models.base import BaseMosaicModel
 from composer.models.classify_mnist.mnist_hparams import MnistClassifierHparams
 from composer.optim.optimizer_hparams import SGDHparams
-from composer.trainer.ddp import DDP
 from composer.trainer.devices.device_gpu import DeviceGPU
 from composer.trainer.trainer import Trainer
 from composer.trainer.trainer_hparams import TrainerHparams
-from composer.utils import ensure_tuple
+from composer.utils import ddp, ensure_tuple
 
 
-def get_total_loss(model: BaseMosaicModel, dataloader: DataLoader, ddp: DDP):
+def get_total_loss(model: BaseMosaicModel, dataloader: DataLoader):
     with torch.no_grad():
         total_loss = 0
         for batch in dataloader:
@@ -27,7 +26,7 @@ def get_total_loss(model: BaseMosaicModel, dataloader: DataLoader, ddp: DDP):
 
         total_loss_tensor = torch.Tensor([total_loss])
         ddp.all_reduce(total_loss_tensor)
-        return total_loss_tensor.item() / ddp.world_size
+        return total_loss_tensor.item() / ddp.get_world_size()
 
 
 def train_model(mosaic_trainer_hparams: TrainerHparams, max_epochs: int = 2, run_loss_check: bool = False):
@@ -69,10 +68,10 @@ def train_model(mosaic_trainer_hparams: TrainerHparams, max_epochs: int = 2, run
         original_model = trainer.device.module_to_device(original_model)
 
     if run_loss_check and trainer.state.train_dataloader:
-        initial_loss = get_total_loss(original_model, trainer.state.train_dataloader, trainer.ddp)
+        initial_loss = get_total_loss(original_model, trainer.state.train_dataloader)
 
         unwrapped_model = trainer.state.model.module
         assert isinstance(unwrapped_model, BaseMosaicModel)
-        post_fit_loss = get_total_loss(unwrapped_model, trainer.state.train_dataloader, trainer.ddp)
+        post_fit_loss = get_total_loss(unwrapped_model, trainer.state.train_dataloader)
 
         assert post_fit_loss < initial_loss + 1e-5
